begin to parse & filter usb midi packets
This splits bulk read buffers into 4 byte midi packets and uses a
basic packet filter to reduce the number of packets that get passed
back up the event stack.

In particular, sequencer playback generates 0f f8 00 00 clock tick
messages at the rate of 24 per quarter note, which is _a lot_
compared to note on/off messages. Filtering clock ticks should free
up a lot of CPU time. If you wanted to make a sequencer that
sync'ed to an external clock, you'd need to actually look at the
clock tick packets. But, that's not relevant for this project.